command_line_args モジュールは削除して subcommand_legacy に統合
